Rewrite
This has essentially completely changed the plugin. It remains for
showing a popup over a flutter map but is now a replacement for a normal
MarkerLayer and is not compatible with the clustering plugin.

My intention is to propose changes directly to the clustering plugin and
leave this as an alternative to a normal MarkerLayer.