Working for both rotated and non-rotated Markers
Snapping relative to the marker now works for both rotated and
non-rotated Markers. The popup itself is not rotated and in the case of
non-rotated markers will snap relative to the center of the Marker.

The latlong dependency has been removed since it is not used directly by
this package.